Осуществила работу поиска
